fixed folder structure
